requirements.txt
Flask==2.3.3
line-bot-sdk==3.5.0
openai==1.3.0
anthropic==0.7.0
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0